fix ios

+++ src/SDLApplication/SDLStdioLogger.h
@@ -14,7 +14,7 @@ namespace Mengine
         ~SDLStdioLogger();
 
     public:
-        void log( EMessageLevel _level, uint32_t _flag, const Char * _data, size_t _count ) override;
+        void log( EMessageLevel _level, uint32_t _flag, const Char * _data, uint32_t _count ) override;
         void flush() override;
     };
 }
